+++ source/source/blender/editors/space_buttons/space_buttons.c
@@ -164,6 +164,9 @@ static void buttons_main_region_draw(const bContext *C, ARegion *ar)
 	case BCONTEXT_EXPORT:
 		contexts[0] = "export";
 		break;
+	case BCONTEXT_INPUT:
+		contexts[0] = "input";
+		break;
 	case BCONTEXT_OBJECT:
 		contexts[0] = "object";
 		break;